# backend/api-docs.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: https://localhost:8443
  description: Generated server url
paths:
  /api/users/{id}:
    get:
      tags:
      - user-rest-controller
      operationId: findUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/FindUserResponseDto'
    put:
      tags:
      - user-rest-controller
      operationId: updateUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequestDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/UpdateUserResponseDto'
...
  /api/users:
    get:
      tags:
      - user-rest-controller
      operationId: findUsers
      parameters:
      - name: page
        in: query
        required: true
        schema:
          type: integer
          format: int32
      - name: amount
        in: query
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FindUserResponseDto'
    post:
      tags:
      - user-rest-controller
      operationId: createUser
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterUserRequestDto'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RegisterUserResponseDto'
  /api/users/{id}/enable:
    post:
      tags:
      - user-rest-controller
      operationId: enableUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/EnableUserResponseDto'
  /api/users/{id}/disable:
    post:
      tags:
      - user-rest-controller
      operationId: disableUser
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/DisableUserResponseDto'
...
  /api/users/{id}/picture:
    get:
      tags:
      - user-rest-controller
      operationId: getImage
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: OK
          content:
            image/jpeg:
              schema:
                type: array
                items:
                  type: string
                  format: byte
            image/png:
              schema:
                type: array
                items:
                  type: string
                  format: byte
...
components:
  schemas:
    UpdateUserRequestDto:
      type: object
      properties:
        encoded_picture:
          type: string
        content_type:
          type: string
        first_name:
          type: string
        surname:
          type: string
        address:
          type: string
        email:
          type: string
        password:
          type: string
    UpdateUserResponseDto:
      type: object
      properties:
        success:
          type: boolean
...
    RegisterUserRequestDto:
      required:
      - address
      - email
      - first_name
      - password
      - surname
      type: object
      properties:
        first_name:
          pattern: "^[a-zA-Z -]+$"
          type: string
        surname:
          pattern: "^[a-zA-Z -]+$"
          type: string
        address:
          type: string
        email:
          maxLength: 255
          minLength: 0
          type: string
        password:
          pattern: "^(?=.*\\d)(?=.*[a-z])(?=.*[A-Z])[0-9a-zA-Z]{8,}$"
          type: string
    RegisterUserResponseDto:
      type: object
      properties:
        success:
          type: boolean
    EnableUserResponseDto:
      type: object
      properties:
        success:
          type: boolean
    DisableUserResponseDto:
      type: object
      properties:
        success:
          type: boolean
...
    FindUserResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        first_name:
          type: string
        surname:
          type: string
        address:
          type: string
        email:
          type: string
        enabled:
          type: boolean
...